feat: add screen for manual token entry
My plan is to use the simulator to test development changes to the CLI
locally. I do not know how to use the camera in the simulator, and how I
could someone take a picture of my terminal. I figured it would be
easier to paste in a string into the app just like Google Authenticator
supports manually adding a secret.